Move implementation guidance to a doc of its own
Separation of concerns: If we want to point driver users to Neo4j::Types for documentation (as we already do), anything that is solely relevant to those who write the drivers themselves is a distraction.

Additionally, the implementation guidance is easier to expand in future if it lives in a document of its own.

This commit does not change the substance of the guidance.