Move POD into files of their own
GitHub will render separate .pod files, which can make the documentation easier to read. This was inspired by @majensen's pattern to create Markdown files and has that same effect, except that it avoids duplication and doesn't modify links to point to GitHub, too.

This commit does not change the module or documentation.