Define reading non-existent property to yield undef
The description of get() already says it returns an undefined value. This change clarifies that returning an empty list is not allowed, which makes get() safe to use from within a hash declaration.

The changed definition implies that there is no way for users of get() to make a distinction between a non-existent property and a property that has the Neo4j null value. To ensure that such a distinction remains possible if needed, this change also adds a note requiring the exists function to give a useful result for fields in the properties() hash.

Tests and the default implementations are changed accordingly.

Resolves #9.