Define methods and behaviour for Neo4j element ID
If the element ID is unavailable for some reason, the legacy numeric ID is to be returned instead. This matches what the official Neo4j drivers tend to do. However, in doing so, there's a risk of users feeding numeric IDs into Cypher's elementId() function, which may cause silent failures. Therefore, Neo4j::Types requires implementors to issue a Perl warning if a numeric ID is given in place of the element ID.