Remove the term "role" from the documentation
In the context of Moose et al., the term "role" has a specific meaning. We diverge quite a bit from that meaning here, which is fine since we have nothing to do with Moose. But the difference in meaning still has potential to cause confusion.

My thinking was that Neo4j::Types might eventually be converted to Moose-style roles, perhaps in version 3 or so. In such a scenario, getting users to think of these modules as roles earlier than that would have been advantageous. But it actually seems rather unlikely this would happen given the backwards compat concerns. Much more likely would be a new set of modules (e. g. Neo4j::Role::Node), which perhaps in turn override isa() to declare that they conform to the Neo4j::Types::Node interface.